Adapt the lounge tests to the shared cleanup harness
Rebasing onto the CI cleanup work left the lounge tests using the pattern it
moved away from, and two incompatibilities that applied cleanly but broke at
runtime.

- Cleanup moves out of per-spec hooks into tests/helpers/lounge.ts, swept by
  global-cleanup like the creation and forum fixtures. Throwaway players are
  tagged per spec so two files cannot see each other's lineups.
- createCircuits/createCup/createMulticup now require an author. Passing none
  sent auteur=undefined, which failed the save and also put the fixtures
  outside the "e2e-" scope that cleanupCreations sweeps, so they leaked.
- Specs sharing global lounge state are merged so serial mode can cover them:
  join.php puts everyone into the tier's one open queue, and lounge_tick()
  finishes every launched queue rather than the caller's.
- Staged matches are inserted as "launching" and published once complete, so a
  tick from an unrelated test cannot finish a half-built match.
