+++ tests/global-cleanup.ts
@@ -1,7 +1,7 @@
 import { FullConfig } from '@playwright/test';
 import { cleanupCreations } from './helpers/mkpc';
 import { cleanupTopics } from './helpers/forum';
-import { cleanupLoungeQueues } from './helpers/lounge';
+import { cleanupLoungeQueues, cleanupLoungeFixtures } from './helpers/lounge';
 
 // Runs once before the suite and once after it, outside every worker.
 //
@@ -22,4 +22,5 @@ export default async function globalCleanup(config: FullConfig) {
   await cleanupCreations(baseURL);
   await cleanupTopics();
   await cleanupLoungeQueues();
+  await cleanupLoungeFixtures();
 }
